+++ portal/src/Navbar.svelte
@@ -1,0 +1,23 @@
+<script>
+</script>
+
+<nav>
+  <header>
+    <a href="#/">SWAPI</a>
+    <label for="nav">(Toggle)</label>
+  </header>
+  <input type="checkbox" id="nav" hidden checked>
+  <div overflow>
+      <ul>
+        <li>
+          <a href="/">Home</a>
+        </li>
+        <li>
+          <a href="https://github.com/rodolphocastro/deno-swapi">Source Code</a>
+        </li>
+        <li>
+          <a href="https://www.linkedin.com/in/rodolpho-alves93/">LinkedIn</a>
+        </li>
+      </ul>
+  </div>
+</nav>